+++ cmd/plugins.go
@@ -7,6 +7,7 @@ import (
 	_ "github.com/tsarna/vinculum/clients/kafka"
 	_ "github.com/tsarna/vinculum/clients/mqtt"
 	_ "github.com/tsarna/vinculum/clients/openai"
+	_ "github.com/tsarna/vinculum/clients/otlp"
 	_ "github.com/tsarna/vinculum/clients/vws"
 	_ "github.com/tsarna/vinculum/servers/http"
 	_ "github.com/tsarna/vinculum/servers/metrics"