+++ config/config.go
@@ -61,6 +61,7 @@ type Config struct {
 	TriggerDefRanges map[string]hcl.Range
 
 	MetricsServers map[string]MetricsRegistrar
+	OtlpClients    map[string]OtlpClient
 }
 
 func NewConfig() *ConfigBuilder {